requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
langchain_google_genai
youtube_transcript_api
pathlib
flask